+++ src/execute.cc
@@ -17,6 +17,7 @@
 
 #include <chrono>
 #include <string.h>
+#include <stdarg.h>
 
 #include "collection.h"
 #include "config.h"
@@ -290,10 +291,12 @@ unwind_stack(Finally_Reason why, Var value, enum outcome *outcome)
                 a->top_rt_stack = new_top;
                 if (found) {
                     *(a->top_rt_stack++) = value;
+                    /* Another exception has been raised... */
                     return 0;
                 }
-            } else
+            } else {
                 free_var(v);
+            }
         }
         if (why == FIN_EXIT) {
             a->pc = value.v.list[2].v.num;
@@ -861,10 +864,16 @@ run(char raise, enum error resumption_error, Var * result)
             STORE_STATE_VARIABLES();                                            \
             if (raise_error(make_raise_pack(the_err, the_msg, the_value), 0)) { \
                 /* No try-except */                                             \
+                free_str(the_msg);                                              \
+                if (the_value.type == TYPE_LIST)                                \
+                    free_var(the_value);                                        \
                 return OUTCOME_ABORTED;                                         \
             } else {                                                            \
                 LOAD_STATE_VARIABLES();                                         \
-                free(the_msg);                                                  \
+                free_str(the_msg);                                              \
+                if (the_value.type == TYPE_LIST) {                              \
+                    free_var(the_value);                                        \
+                }                                                               \
                 goto next_opcode;                                               \
             } \
         } \
@@ -879,10 +888,10 @@ run(char raise, enum error resumption_error, Var * result)
     } while (0)
 
     /* NOTE: the_msg will be freed */
-#define PUSH_RAISED_ERROR(the_err, the_msg, the_value)                            \
+#define PUSH_ERROR_WITH_VALUE(the_err, the_msg, the_value)                        \
     do {                                                                          \
         RAISE_ERROR_WITH_VALUE(the_err, the_msg, the_value); /* may not return */ \
-        free(the_msg);                                                            \
+        free_str(the_msg);                                                        \
         free_var(the_value);                                                      \
         error_var.type = TYPE_ERR;                                                \
         error_var.v.err = the_err;                                                \
@@ -904,16 +913,33 @@ run(char raise, enum error resumption_error, Var * result)
             PUSH_ERROR(the_err);                                        \
     } while (0)
 
-#define RAISE_X_NOT_FOUND(the_err, the_missing)                               \
-    do {                                                                      \
-        Var missing;                                                          \
-        missing.type = TYPE_STR;                                              \
-        missing.v.str = str_dup(the_missing);                                 \
-        char *error_msg = nullptr;                                            \
-        asprintf(&error_msg, "%s: %s", unparse_error(the_err), the_missing);  \
-        PUSH_RAISED_ERROR(the_err, error_msg, missing);                       \
+#define PUSH_X_NOT_FOUND(the_err, the_missing)                                      \
+    do {                                                                            \
+        Var missing = str_dup_to_var(the_missing);                                  \
+        static Stream *error_stream = nullptr;                                      \
+        if (error_stream == nullptr) {                                              \
+            error_stream = new_stream(20);                                          \
+        }                                                                           \
+        stream_printf(error_stream, "%s: %s", unparse_error(the_err), the_missing); \
+        char *error_message = str_dup(reset_stream(error_stream));                  \
+        PUSH_ERROR_WITH_VALUE(the_err, error_message, missing);                     \
     } while (0)
 
+#define RAISE_TYPE_MISMATCH(...)                               \
+    do {                                                       \
+        char *error_msg = type_mismatch_string(__VA_ARGS__);   \
+        Var value = type_mismatch_value(__VA_ARGS__);          \
+        RAISE_ERROR_WITH_VALUE(E_TYPE, error_msg, value);      \
+        free_str(error_msg);                                   \
+        free_var(value);                                       \
+    } while (0)
+
+#define PUSH_TYPE_MISMATCH(...)                                \
+    do {                                                       \
+        char *error_msg = type_mismatch_string(__VA_ARGS__);   \
+        Var value = type_mismatch_value(__VA_ARGS__);          \
+        PUSH_ERROR_WITH_VALUE(E_TYPE, error_msg, value);       \
+    } while (0)
 
 #define JUMP(label)     (bv = bc.vector + label)
 
@@ -979,9 +1005,12 @@ do_test:
 
                 to = TOP_RT_VALUE;
                 from = NEXT_TOP_RT_VALUE;
+
                 if ((to.type != TYPE_INT && to.type != TYPE_OBJ)
                         || to.type != from.type) {
-                    RAISE_ERROR(E_TYPE);
+                    RAISE_TYPE_MISMATCH(2,
+                                        to.type != TYPE_INT && to.type != TYPE_OBJ ? to.type : from.type,
+                                        TYPE_INT, TYPE_OBJ);
                     free_var(POP());
                     free_var(POP());
                     JUMP(lab);
@@ -1055,7 +1084,7 @@ do_test:
                     free_var(key);
                     free_var(value);
                     free_var(map);
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(8, key.type, TYPE_STR, TYPE_INT, TYPE_OBJ, TYPE_ERR, TYPE_FLOAT, TYPE_ANON, TYPE_WAIF, TYPE_BOOL);
                 } else {
                     r = mapinsert(map, key, value);
                     if (value_bytes(r) <= server_int_option_cached(SVO_MAX_MAP_VALUE_BYTES))
@@ -1086,7 +1115,7 @@ do_test:
                 if (list.type != TYPE_LIST) {
                     free_var(list);
                     free_var(tail);
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(1, list.type, TYPE_LIST);
                 } else {
                     r = listappend(list, tail);
                     if (value_bytes(r) <= server_int_option_cached(SVO_MAX_LIST_VALUE_BYTES))
@@ -1108,7 +1137,7 @@ do_test:
                 if (tail.type != TYPE_LIST || list.type != TYPE_LIST) {
                     free_var(list);
                     free_var(tail);
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(1, tail.type != TYPE_LIST ? tail.type : list.type, TYPE_LIST);
                 } else {
                     r = listconcat(list, tail);
                     if (value_bytes(r) <= server_int_option_cached(SVO_MAX_LIST_VALUE_BYTES))
@@ -1229,8 +1258,9 @@ do_test:
 
             case OP_CHECK_LIST_FOR_SPLICE:
                 if (TOP_RT_VALUE.type != TYPE_LIST) {
+                    var_type rt_value_type = TOP_RT_VALUE.type;
                     free_var(POP());
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(1, rt_value_type, TYPE_LIST);
                 }
                 /* no op if top-rt-stack is a list */
                 break;
@@ -1285,7 +1315,7 @@ do_test:
                 } else if (rhs.type != lhs.type || rhs.type == TYPE_LIST || rhs.type == TYPE_MAP) {
                     free_var(rhs);
                     free_var(lhs);
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(1, rhs.type, lhs.type);
                 } else {
                     switch (rhs.type) {
                         case TYPE_INT:
@@ -1345,9 +1375,11 @@ finish_comparison:
                     free_var(lhs);
                     free_var(rhs);
                 } else if (rhs.type != TYPE_LIST && rhs.type != TYPE_MAP) {
+                    var_type rhs_type = rhs.type;
+                    var_type lhs_type = lhs.type;
                     free_var(rhs);
                     free_var(lhs);
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(2, rhs.type, TYPE_LIST, TYPE_MAP);
                 } else {
                     ans.type = TYPE_INT;
                     ans.v.num = ismember(lhs, rhs, 0);
@@ -1364,6 +1396,7 @@ finish_comparison:
             case OP_MOD:
             {
                 Var lhs, rhs, ans;
+                var_type lhs_type, rhs_type;
 
                 rhs = POP();    /* should be number */
                 lhs = POP();    /* should be number */
@@ -1389,13 +1422,21 @@ finish_comparison:
                 } else {
                     ans.type = TYPE_ERR;
                     ans.v.err = E_TYPE;
+                    lhs_type = lhs.type;
+                    rhs_type = rhs.type;
                 }
                 free_var(rhs);
                 free_var(lhs);
-                if (ans.type == TYPE_ERR)
-                    PUSH_ERROR(ans.v.err);
-                else
+                if (ans.type == TYPE_ERR) {
+                    if (ans.v.err == E_TYPE)
+                        PUSH_TYPE_MISMATCH(2,
+                                           lhs_type != TYPE_INT && lhs_type != TYPE_FLOAT ? lhs_type : rhs_type,
+                                           TYPE_INT, TYPE_FLOAT);
+                    else
+                        PUSH_ERROR(ans.v.err);
+                } else {
                     PUSH(ans);
+                }
             }
             break;
 
@@ -1405,6 +1446,10 @@ finish_comparison:
 
                 rhs = POP();
                 lhs = POP();
+
+                var_type lhs_type = lhs.type;
+                var_type rhs_type = rhs.type;
+
                 if ((lhs.type == TYPE_INT || lhs.type == TYPE_FLOAT)
                         && (rhs.type == TYPE_INT || rhs.type == TYPE_FLOAT))
                     ans = do_add(lhs, rhs);
@@ -1431,10 +1476,27 @@ finish_comparison:
                 free_var(rhs);
                 free_var(lhs);
 
-                if (ans.type == TYPE_ERR)
-                    PUSH_ERROR_UNLESS_QUOTA(ans.v.err);
-                else
+                if (ans.type == TYPE_ERR) {
+                    /* This is maybe slightly unwieldy, but it works. It assumes lhs is the type you really wanted. */
+                    if (ans.v.err == E_TYPE) {
+                        if (lhs_type == TYPE_STR)
+                            PUSH_TYPE_MISMATCH(1, lhs_type != TYPE_STR ? lhs_type : rhs_type, TYPE_STR);
+                        else if (rhs_type == TYPE_STR && lhs_type != TYPE_INT && lhs_type != TYPE_FLOAT)
+                            PUSH_TYPE_MISMATCH(1, lhs_type, TYPE_STR);
+                        else if (lhs_type == TYPE_INT)
+                            PUSH_TYPE_MISMATCH(1, rhs_type, TYPE_INT);
+                        else if (lhs_type == TYPE_FLOAT)
+                            PUSH_TYPE_MISMATCH(1, rhs_type, TYPE_FLOAT);
+                        else if (lhs_type != TYPE_INT && lhs_type != TYPE_FLOAT && (rhs_type == TYPE_INT || rhs_type == TYPE_FLOAT))
+                            PUSH_TYPE_MISMATCH(1, lhs_type, rhs.type);
+                        else
+                            PUSH_TYPE_MISMATCH(3, lhs.type, TYPE_INT, TYPE_FLOAT, TYPE_STR);
+                    } else {
+                        PUSH_ERROR_UNLESS_QUOTA(ans.v.err);
+                    }
+                } else {
                     PUSH(ans);
+                }
             }
             break;
 
@@ -1469,6 +1531,7 @@ finish_comparison:
             case OP_UNARY_MINUS:
             {
                 Var arg, ans;
+                var_type arg_type;
 
                 arg = POP();
                 if (arg.type == TYPE_INT) {
@@ -1478,8 +1541,9 @@ finish_comparison:
                     ans.type = TYPE_FLOAT;
                     ans.v.fnum = -arg.v.fnum;
                 } else {
+                    arg_type = arg.type;
                     free_var(arg);
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(2, arg_type, TYPE_INT, TYPE_FLOAT);
                     break;
                 }
 
@@ -1595,12 +1659,11 @@ finish_comparison:
 
                 index = TOP_RT_VALUE;
                 list = NEXT_TOP_RT_VALUE;
-
                 if (list.type == TYPE_MAP) {
                     Var value;
                     const rbnode *node;
                     if (index.is_collection() && TYPE_ANON != index.type) {
-                        PUSH_ERROR(E_TYPE);
+                        PUSH_TYPE_MISMATCH(8, index.type, TYPE_STR, TYPE_INT, TYPE_OBJ, TYPE_ERR, TYPE_FLOAT, TYPE_ANON, TYPE_WAIF, TYPE_BOOL);
                     } else if (!(node = maplookup(list, index, &value, 0))) {
                         PUSH_ERROR(E_RANGE);
                     } else {
@@ -1609,7 +1672,7 @@ finish_comparison:
                     }
                 } else if (list.type == TYPE_LIST) {
                     if (index.type != TYPE_INT) {
-                        PUSH_ERROR(E_TYPE);
+                        PUSH_TYPE_MISMATCH(1, index.type, TYPE_INT);
                     } else if (index.v.num <= 0 ||
                                index.v.num > list.v.list[0].v.num) {
                         PUSH_ERROR(E_RANGE);
@@ -1618,7 +1681,7 @@ finish_comparison:
                         list.v.list[index.v.num].type = TYPE_NONE;
                     }
                 } else {
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(2, list.type, TYPE_LIST, TYPE_MAP);
                 }
             }
             break;
@@ -1633,22 +1696,25 @@ finish_comparison:
 
                 if (base.type != TYPE_MAP && base.type != TYPE_LIST
                         && base.type != TYPE_STR) {
+                    var_type base_type = base.type;
                     free_var(to);
                     free_var(from);
                     free_var(base);
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(3, base_type, TYPE_LIST, TYPE_MAP, TYPE_STR);
                 } else if (base.type == TYPE_MAP
                            && ((to.is_collection() && TYPE_ANON != to.type) || (from.is_collection() && TYPE_ANON != from.type))) {
+                    var_type to_from_type = to.is_collection() ? to.type : from.type;
                     free_var(to);
                     free_var(from);
                     free_var(base);
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(8, to_from_type, TYPE_STR, TYPE_INT, TYPE_OBJ, TYPE_ERR, TYPE_FLOAT, TYPE_ANON, TYPE_WAIF, TYPE_BOOL);
                 } else if ((base.type == TYPE_LIST || base.type == TYPE_STR)
                            && (to.type != TYPE_INT || from.type != TYPE_INT)) {
+                    var_type to_from_type = to.type != TYPE_INT ? to.type : from.type;
                     free_var(to);
                     free_var(from);
                     free_var(base);
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(1, to_from_type, TYPE_INT);
                 } else if (base.type == TYPE_MAP) {
                     Var iterfrom, iterto;
                     int rel = compare(from, to, 0);
@@ -1712,7 +1778,7 @@ finish_comparison:
                 int var_pos = READ_BYTES(bv, bc.numbytes_var_name);
                 value = RUN_ACTIV.rt_env[var_pos];
                 if (value.type == TYPE_NONE) {
-                    RAISE_X_NOT_FOUND(E_VARNF, *(&RUN_ACTIV.prog->var_names[var_pos]));
+                    PUSH_X_NOT_FOUND(E_VARNF, *(&RUN_ACTIV.prog->var_names[var_pos]));
                 }
                 else
                     PUSH_REF(value);
@@ -1736,9 +1802,10 @@ finish_comparison:
                     else
                         PUSH_ERROR(err);
                 } else if (!obj.is_object() || propname.type != TYPE_STR) {
+                    var_type incorrect_type = propname.type != TYPE_STR ? propname.type : obj.type;
                     free_var(propname);
                     free_var(obj);
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(1, incorrect_type, incorrect_type == TYPE_STR ? TYPE_OBJ : TYPE_STR);
                 } else if (!is_valid(obj)) {
                     free_var(propname);
                     free_var(obj);
@@ -1752,9 +1819,11 @@ finish_comparison:
 
                     free_var(obj);
 
-                    if (!h.ptr)
-                        RAISE_X_NOT_FOUND(E_PROPNF, propname.v.str);
-                    else {
+                    if (!h.ptr) {
+                        const char* prop_name = propname.v.str;
+                        free_var(propname);
+                        PUSH_X_NOT_FOUND(E_PROPNF, prop_name);
+                    } else {
                         free_var(propname);
 
                         if (built_in
@@ -1784,9 +1853,10 @@ finish_comparison:
                         PUSH(prop);
                     else
                         PUSH_ERROR(err);
-                } else if (!obj.is_object() || propname.type != TYPE_STR)
-                    PUSH_ERROR(E_TYPE);
-                else if (!is_valid(obj))
+                } else if (!obj.is_object() || propname.type != TYPE_STR) {
+                    var_type incorrect_type = propname.type != TYPE_STR ? propname.type : obj.type;
+                    PUSH_TYPE_MISMATCH(1, incorrect_type, incorrect_type == TYPE_STR ? TYPE_OBJ : TYPE_STR);
+                } else if (!is_valid(obj))
                     PUSH_ERROR(E_INVIND);
                 else {
                     db_prop_handle h;
@@ -1795,7 +1865,7 @@ finish_comparison:
                     h = db_find_property(obj, propname.v.str, &prop);
                     built_in = db_is_property_built_in(h);
                     if (!h.ptr)
-                        RAISE_X_NOT_FOUND(E_PROPNF, propname.v.str);
+                        PUSH_X_NOT_FOUND(E_PROPNF, propname.v.str);
                     else if (built_in
                              ? bi_prop_protected(built_in, RUN_ACTIV.progr)
                              : !db_property_allows(h, RUN_ACTIV.progr, PF_READ))
@@ -1826,17 +1896,18 @@ finish_comparison:
                     } else {
                         free_var(rhs);
                         if (err == E_PROPNF)
-                            RAISE_X_NOT_FOUND(E_PROPNF, propname.v.str);
+                            PUSH_X_NOT_FOUND(E_PROPNF, propname.v.str);
                         else {
                             free_var(propname);
                             PUSH_ERROR(err);
                         }
                     }
                 } else if (!obj.is_object() || propname.type != TYPE_STR) {
+                    var_type incorrect_type = propname.type != TYPE_STR ? propname.type : obj.type;
                     free_var(rhs);
                     free_var(propname);
                     free_var(obj);
-                    PUSH_ERROR(E_TYPE);
+                    PUSH_TYPE_MISMATCH(1, incorrect_type, incorrect_type == TYPE_STR ? TYPE_OBJ : TYPE_STR);
                 } else if (!is_valid(obj)) {
                     free_var(rhs);
                     free_var(propname);
@@ -1917,7 +1988,9 @@ finish_comparison:
                     free_var(obj);
 
                     if (err == E_PROPNF) {
-                        RAISE_X_NOT_FOUND(E_PROPNF, propname.v.str);
+                        const char *prop_name = propname.v.str;
+                        free_var(propname);
+                        PUSH_X_NOT_FOUND(E_PROPNF, prop_name);
                     } else {
                         free_var(propname);
                         if (err == E_NONE) {
@@ -1944,8 +2017,9 @@ finish_comparison:
                 if (op == OP_FORK_WITH_ID)
                     id = READ_BYTES(bv, bc.numbytes_var_name);
                 if (time.type != TYPE_INT && time.type != TYPE_FLOAT) {
+                    var_type time_type = time.type;
                     free_var(time);
-                    RAISE_ERROR(E_TYPE);
+                    RAISE_TYPE_MISMATCH(2, time_type, TYPE_INT, TYPE_FLOAT);
                 }
                 when = time.type == TYPE_INT ? time.v.num : time.v.fnum;
                 free_var(time);
@@ -2032,17 +2106,19 @@ else if (obj.type == TYPE_##t1) {           \
                 }
                 free_var(obj);
 
-                if (err == E_VERBNF)
-                    RAISE_X_NOT_FOUND(E_VERBNF, verb.v.str);
-                else {
-                    free_var(verb);
-
-                    if (err != E_NONE) {
-                        /* there is an error, RUN_ACTIV unchanged,
-                                       args must be freed */
-                        free_var(args);
+                if (err != E_NONE) {
+                    /* there is an error, RUN_ACTIV unchanged, args must be freed */
+                    free_var(args);
+                    if (err == E_VERBNF) {
+                        const char *verb_name = verb.v.str;
+                        free_var(verb);
+                        PUSH_X_NOT_FOUND(err, verb_name);
+                    } else {
+                        free_var(verb);
                         PUSH_ERROR(err);
                     }
+                } else {
+                    free_var(verb);
                 }
             }
             break;
@@ -2148,25 +2224,32 @@ else if (obj.type == TYPE_##t1) {           \
                         if ((base.type != TYPE_MAP && base.type != TYPE_LIST
                                 && base.type != TYPE_STR)
                                 || (base.type != value.type)) {
+                            var_type base_type = base.type;
+                            var_type value_type = value.type;
                             free_var(to);
                             free_var(from);
                             free_var(base);
                             free_var(value);
-                            PUSH_ERROR(E_TYPE);
+                            if (base_type != value_type)
+                                PUSH_TYPE_MISMATCH(1, value_type, base.type);
+                            else
+                                PUSH_TYPE_MISMATCH(3, base_type, TYPE_LIST, TYPE_MAP, TYPE_STR);
                         } else if (base.type == TYPE_MAP
                                    && ((to.is_collection() && TYPE_ANON != to.type) || (from.is_collection() && TYPE_ANON != from.type))) {
+                            var_type to_from_type = to.is_collection() ? to.type : from.type;
                             free_var(to);
                             free_var(from);
                             free_var(base);
                             free_var(value);
-                            PUSH_ERROR(E_TYPE);
+                            PUSH_TYPE_MISMATCH(8, to_from_type, TYPE_STR, TYPE_INT, TYPE_OBJ, TYPE_ERR, TYPE_FLOAT, TYPE_ANON, TYPE_WAIF, TYPE_BOOL);
                         } else if ((base.type == TYPE_LIST || base.type == TYPE_STR)
                                    && (to.type != TYPE_INT || from.type != TYPE_INT)) {
+                            var_type to_from_type = to.type != TYPE_INT ? to.type : from.type;
                             free_var(to);
                             free_var(from);
                             free_var(base);
                             free_var(value);
-                            PUSH_ERROR(E_TYPE);
+                            PUSH_TYPE_MISMATCH(1, to_from_type, TYPE_INT);
                         } else if (base.type == TYPE_MAP) {
                             Var res = none;
                             Var iterfrom, iterto;
@@ -2252,7 +2335,7 @@ else if (obj.type == TYPE_##t1) {           \
                                 : var_ref(none);
                             PUSH(v);
                         } else
-                            PUSH_ERROR(E_TYPE);
+                            PUSH_TYPE_MISMATCH(3, item.type, TYPE_STR, TYPE_LIST, TYPE_MAP);
                     }
                     break;
 
@@ -2277,7 +2360,7 @@ else if (obj.type == TYPE_##t1) {           \
                                 : var_ref(none);
                             PUSH(v);
                         } else
-                            PUSH_ERROR(E_TYPE);
+                            PUSH_TYPE_MISMATCH(3, item.type, TYPE_STR, TYPE_LIST, TYPE_MAP);
                     }
                     break;
 
@@ -2317,7 +2400,10 @@ else if (obj.type == TYPE_##t1) {           \
 
                         if (e != E_NONE) {  /* skip rest of operands */
                             free_var(POP());    /* replace list with error code */
-                            PUSH_ERROR(e);
+                            if (e == E_TYPE)
+                                PUSH_TYPE_MISMATCH(1, list.type, TYPE_LIST);
+                            else
+                                PUSH_ERROR(e);
                             for (i = 1; i <= nargs; i++) {
                                 SKIP_BYTES(bv, bc.numbytes_var_name);
                                 SKIP_BYTES(bv, bc.numbytes_label);
@@ -2396,8 +2482,10 @@ else if (obj.type == TYPE_##t1) {           \
                             v = POP();
 
                         marker = POP();
-                        if (marker.type != TYPE_CATCH)
+                        if (marker.type != TYPE_CATCH) {
+                            errlog("Marker is type %i with value %i\n", marker.type, marker.v.num);
                             panic_moo("Stack marker is not TYPE_CATCH!");
+                        }
                         for (i = 0; i < marker.v.num; i++) {
                             (void) POP();   /* handler PC */
                             free_var(POP());    /* code list */
@@ -2486,7 +2574,7 @@ else if (obj.type == TYPE_##t1) {           \
 
                         if (BASE.type != TYPE_STR && BASE.type != TYPE_LIST
                                 && BASE.type != TYPE_MAP) {
-                            RAISE_ERROR(E_TYPE);
+                            RAISE_TYPE_MISMATCH(3, BASE.type, TYPE_STR, TYPE_LIST, TYPE_MAP);
                             free_var(POP());
                             free_var(POP());
                             JUMP(lab);
@@ -2548,7 +2636,7 @@ else if (obj.type == TYPE_##t1) {           \
 
                         if (BASE.type != TYPE_STR && BASE.type != TYPE_LIST
                                 && BASE.type != TYPE_MAP) {
-                            RAISE_ERROR(E_TYPE);
+                            RAISE_TYPE_MISMATCH(3, BASE.type, TYPE_STR, TYPE_LIST, TYPE_MAP);
                             free_var(POP());
                             free_var(POP());
                             JUMP(lab);
@@ -2606,6 +2694,7 @@ else if (obj.type == TYPE_##t1) {           \
                     case EOP_BITXOR:
                     {
                         Var rhs, lhs, ans;
+                        var_type lhs_type, rhs_type;
 
                         rhs = POP();
                         lhs = POP();
@@ -2622,12 +2711,17 @@ else if (obj.type == TYPE_##t1) {           \
                         } else {
                             ans.type = TYPE_ERR;
                             ans.v.err = E_TYPE;
+                            lhs_type = lhs.type;
+                            rhs_type = rhs.type;
                         }
 
                         free_var(lhs);
                         free_var(rhs);
                         if (ans.type == TYPE_ERR)
-                            PUSH_ERROR(ans.v.err);
+                            if (ans.v.err == E_TYPE)
+                                PUSH_TYPE_MISMATCH(1, lhs_type != TYPE_INT ? lhs_type : rhs_type, TYPE_INT);
+                            else
+                                PUSH_ERROR(ans.v.err);
                         else
                             PUSH(ans);
                     }
@@ -2637,12 +2731,15 @@ else if (obj.type == TYPE_##t1) {           \
                     case EOP_BITSHR:
                     {
                         Var rhs, lhs, ans;
+                        var_type lhs_type, rhs_type;
 
                         rhs = POP();
                         lhs = POP();
                         if (lhs.type != TYPE_INT || rhs.type != TYPE_INT) {
                             ans.type = TYPE_ERR;
                             ans.v.err = E_TYPE;
+                            lhs_type = lhs.type;
+                            rhs_type = rhs.type;
                         } else if (rhs.v.num > sizeof(Num) * CHAR_BIT || rhs.v.num < 0) {
                             ans.type = TYPE_ERR;
                             ans.v.err = E_INVARG;
@@ -2665,7 +2762,10 @@ else if (obj.type == TYPE_##t1) {           \
                         free_var(lhs);
                         free_var(rhs);
                         if (ans.type == TYPE_ERR)
-                            PUSH_ERROR(ans.v.err);
+                            if (ans.v.err == E_TYPE)
+                                PUSH_TYPE_MISMATCH(1, lhs_type != TYPE_INT ? lhs_type : rhs_type, TYPE_INT);
+                            else
+                                PUSH_ERROR(ans.v.err);
                         else
                             PUSH(ans);
                     }
@@ -2674,6 +2774,7 @@ else if (obj.type == TYPE_##t1) {           \
                     case EOP_COMPLEMENT:
                     {
                         Var arg, ans;
+                        var_type arg_type;
 
                         arg = POP();
                         if (arg.type == TYPE_INT) {
@@ -2682,11 +2783,15 @@ else if (obj.type == TYPE_##t1) {           \
                         } else {
                             ans.type = TYPE_ERR;
                             ans.v.err = E_TYPE;
+                            arg_type = arg.type;
                         }
 
                         free_var(arg);
                         if (ans.type == TYPE_ERR)
-                            PUSH_ERROR(ans.v.err);
+                            if (ans.v.err == E_TYPE)
+                                PUSH_TYPE_MISMATCH(1, arg_type, TYPE_INT);
+                            else
+                                PUSH_ERROR(ans.v.err);
                         else
                             PUSH(ans);
                     }
@@ -2744,7 +2849,7 @@ else if (obj.type == TYPE_##t1) {           \
                 value = RUN_ACTIV.rt_env[PUSH_n_INDEX(op)];
                 if (value.type == TYPE_NONE) {
                     free_var(value);
-                    RAISE_X_NOT_FOUND(E_VARNF, *(&RUN_ACTIV.prog->var_names[PUSH_n_INDEX(op)]));
+                    PUSH_X_NOT_FOUND(E_VARNF, *(&RUN_ACTIV.prog->var_names[PUSH_n_INDEX(op)]));
                 } else
                     PUSH_REF(value);
             }
@@ -2787,7 +2892,7 @@ else if (obj.type == TYPE_##t1) {           \
                 Var *vp;
                 vp = &RUN_ACTIV.rt_env[PUSH_CLEAR_n_INDEX(op)];
                 if (vp->type == TYPE_NONE) {
-                    RAISE_X_NOT_FOUND(E_VARNF, *(&RUN_ACTIV.prog->var_names[PUSH_CLEAR_n_INDEX(op)]));
+                    PUSH_X_NOT_FOUND(E_VARNF, *(&RUN_ACTIV.prog->var_names[PUSH_CLEAR_n_INDEX(op)]));
                 } else {
                     PUSH(*vp);
                     vp->type = TYPE_NONE;
@@ -3479,7 +3584,7 @@ bf_pass(Var arglist, Byte next, void *vdata, Objid progr)
     free_var(arglist);
 
     if (e == E_VERBNF)
-        return make_raise_x_not_found_pack(e, RUN_ACTIV.verb);
+        return make_x_not_found_pack(e, RUN_ACTIV.verb);
     else
         return make_error_pack(e);
 }
@@ -3895,4 +4000,64 @@ bool get_thread_mode()
 void set_thread_mode(bool mode)
 {
     RUN_ACTIV.threaded = mode;
+}
+
+/* Create a type mismatch traceback and try-except return value.
+   The first argument is the mismatched type that was received.
+   The remaining arguments are the types that were expected. */
+char*
+type_mismatch_string(int n_args, ...)
+{
+    static Stream *error_msg = nullptr;
+    if (error_msg == nullptr)
+        error_msg = new_stream(40);
+
+    stream_printf(error_msg, "%s (expected", unparse_error(E_TYPE));
+    var_type the_var = TYPE_CLEAR;
+    va_list args;
+    va_start(args, n_args);
+
+    var_type mismatch = (var_type)va_arg(args, int);
+
+    for (int x = 1; x <= n_args; x++) {
+        var_type the_var = (var_type)va_arg(args, int);
+        stream_printf(error_msg, " %s,", parse_type((var_type)the_var));
+        if (x == n_args - 1) {
+            stream_delete_char(error_msg);
+            stream_add_string(error_msg, " or");
+        }
+    }
+    va_end(args);
+
+    stream_delete_char(error_msg); // delete the comma
+    stream_printf(error_msg, "; got %s)", parse_type(mismatch));
+
+    return str_dup(reset_stream(error_msg));
+}
+
+Var
+type_mismatch_value(int n_args, ...)
+{
+    /* Since -d verbs can't use the value anyway,
+       we won't waste time and money setting it up. */
+    if (!RUN_ACTIV.debug)
+        return var_ref(zero);
+
+    Var value_var = new_list(2);
+    Var expected_types = new_list(n_args);
+    va_list args;
+    va_start(args, n_args);
+
+    var_type mismatch = (var_type)va_arg(args, int);
+
+    for (int x = 1; x <= n_args; x++) {
+        var_type the_var = (var_type)va_arg(args, int);
+        expected_types.v.list[x] = Var::new_int(the_var);
+    }
+    va_end(args);
+
+    value_var.v.list[1] = expected_types;
+    value_var.v.list[2] = Var::new_int(mismatch & TYPE_DB_MASK);
+
+    return var_ref(value_var);
 }