+++ src/include/unparse.h
@@ -40,5 +40,6 @@ extern const char *error_name(enum error);	/* E_NONE -> "E_NONE" */
 extern const char *unparse_error(enum error);	/* E_NONE -> "No error" */
 
 extern int parse_error(const char *error);	/* "E_NONE" -> E_NONE */
+extern const char* parse_type(var_type);
 
 #endif